+++ hyperdbg/include/SDK/headers/Constants.h
@@ -17,7 +17,7 @@
 //////////////////////////////////////////////////
 
 #define VERSION_MAJOR 0
-#define VERSION_MINOR 11
+#define VERSION_MINOR 12
 #define VERSION_PATCH 0
 
 //