+++ hyperdbg/libhyperdbg/code/debugger/commands/extension-commands/pcitree.cpp
@@ -26,10 +26,10 @@ CommandPcitreeHelp()
 {
     ShowMessages("!pcitree : enumerates all PCIe endpoints on the debuggee.\n\n");
 
-    ShowMessages("syntax : \t!pcietree\n");
+    ShowMessages("syntax : \t!pcitree\n");
 
     ShowMessages("\n");
-    ShowMessages("\t\te.g : !pcietree\n");
+    ShowMessages("\t\te.g : !pcitree\n");
 }
 
 /**
@@ -56,21 +56,15 @@ CommandPcitree(vector<CommandToken> CommandTokens, string Command)
     }
 
     //
-    // Prepare the buffer
+    // Send buffer
     //
-
     if (g_IsSerialConnectedToRemoteDebuggee)
     {
-        //
-        // Send the request over serial kernel debugger
-        //
         KdSendPcitreePacketToDebuggee(&PcitreePacket);
     }
     else
     {
         AssertShowMessageReturnStmt(g_DeviceHandle, ASSERT_MESSAGE_DRIVER_NOT_LOADED, AssertReturn);
-
-        ShowMessages("Sending IOCTL\n");
 
         //
         // Send IOCTL
@@ -87,8 +81,6 @@ CommandPcitree(vector<CommandToken> CommandTokens, string Command)
             NULL                                             // synchronous call
         );
 
-        ShowMessages("Done sending IOCTL\n");
-
         if (!Status)
         {
             ShowMessages("ioctl failed with code 0x%x\n", GetLastError());
@@ -98,17 +90,49 @@ CommandPcitree(vector<CommandToken> CommandTokens, string Command)
         if (PcitreePacket.KernelStatus == DEBUGGER_OPERATION_WAS_SUCCESSFUL)
         {
             //
-            // Show the results
+            // Print PCI device tree
             //
-            ShowMessages("Result: (0000:00:00) VID:DID: %x:%x\n", PcitreePacket.PciTree.Domain[0].Bus[0].Device[0].ConfigSpace->CommonHeader.VendorId, PcitreePacket.PciTree.Domain[0].Bus[0].Device[0].ConfigSpace->CommonHeader.DeviceId);
+            ShowMessages("%-12s | %-9s | %-17s | %s \n%s\n", "DBDF", "VID:DID", "Vendor Name", "Device Name", "----------------------------------------------------------------------");
+            for (UINT8 i = 0; i < (PcitreePacket.EndpointsTotalNum < EP_MAX_NUM ? PcitreePacket.EndpointsTotalNum : EP_MAX_NUM); i++)
+            {
+                Vendor * CurrentVendor     = GetVendorById(PcitreePacket.Endpoints[i].ConfigSpace.VendorId);
+                CHAR *   CurrentVendorName = (CHAR *)"N/A";
+                CHAR *   CurrentDeviceName = (CHAR *)"N/A";
+
+                if (CurrentVendor != NULL)
+                {
+                    CurrentVendorName      = CurrentVendor->VendorName;
+                    Device * CurrentDevice = GetDeviceFromVendor(CurrentVendor, PcitreePacket.Endpoints[i].ConfigSpace.DeviceId);
+
+                    if (CurrentDevice != NULL)
+                    {
+                        CurrentDeviceName = CurrentDevice->DeviceName;
+                    }
+                }
+
+                ShowMessages("%04x:%02x:%02x:%x | %04x:%04x | %-17.*s | %.*s\n",
+                             0, // TODO: Add support for domains beyond 0000
+                             PcitreePacket.Endpoints[i].Bus,
+                             PcitreePacket.Endpoints[i].Device,
+                             PcitreePacket.Endpoints[i].Function,
+                             PcitreePacket.Endpoints[i].ConfigSpace.VendorId,
+                             PcitreePacket.Endpoints[i].ConfigSpace.DeviceId,
+                             strnlen_s(CurrentVendorName, PCI_NAME_STR_LENGTH),
+                             CurrentVendorName,
+                             strnlen_s(CurrentDeviceName, PCI_NAME_STR_LENGTH),
+                             CurrentDeviceName
+
+                );
+
+                FreeVendor(CurrentVendor);
+            }
+            FreePciIdDatabase();
         }
         else
         {
             //
             // An err occurred, no results
             //
-            ShowMessages("An err occured, no results:");
-
             ShowErrorMessage(PcitreePacket.KernelStatus);
         }
     }