+++ hyperdbg/libhyperdbg/code/debugger/core/interpreter.cpp
@@ -1607,6 +1607,8 @@ InitializeCommandsDictionary()
     g_CommandsList["!pcitree"]  = {&CommandPcitree, &CommandPcitreeHelp, DEBUGGER_COMMAND_PCITREE_ATTRIBUTES};
     g_CommandsList["!pcietree"] = {&CommandPcitree, &CommandPcitreeHelp, DEBUGGER_COMMAND_PCITREE_ATTRIBUTES};
 
+    g_CommandsList["!idt"] = {&CommandIdt, &CommandIdtHelp, DEBUGGER_COMMAND_IDT_ATTRIBUTES};
+
     //
     // hwdbg commands
     //