+++ hyperdbg/libhyperdbg/header/commands.h
@@ -459,6 +459,9 @@ typedef std::map<std::string, COMMAND_DETAIL> CommandType;
 #define DEBUGGER_COMMAND_PCITREE_ATTRIBUTES \
     DEBUGGER_COMMAND_ATTRIBUTE_LOCAL_COMMAND_IN_DEBUGGER_MODE
 
+#define DEBUGGER_COMMAND_IDT_ATTRIBUTES \
+    DEBUGGER_COMMAND_ATTRIBUTE_LOCAL_COMMAND_IN_DEBUGGER_MODE
+
 //////////////////////////////////////////////////
 //             Command Functions                //
 //////////////////////////////////////////////////
@@ -732,6 +735,9 @@ CommandAssemble(vector<CommandToken> CommandTokens, string Command);
 
 VOID
 CommandPcitree(vector<CommandToken> CommandTokens, string Command);
+
+VOID
+CommandIdt(vector<CommandToken> CommandTokens, string Command);
 
 //
 // hwdbg commands